+++ assets/js/index.js
@@ -10,11 +10,6 @@ require("jquery");
 require("bootstrap");
 require("bootstrap-icons/font/bootstrap-icons.css");
 require("bootstrap4-toggle");
-require("darkreader");
-const hljs = require("highlight.js/lib/core");
-const hljsPython = require("highlight.js/lib/languages/python");
-const hljsShell = require("highlight.js/lib/languages/shell");
-const hljsYaml = require("highlight.js/lib/languages/yaml");
 const $ = require("jquery");
 const {
   setMessageCount,
@@ -32,10 +27,6 @@ require("../xml/browserconfig.xml");
 const darkMode = new DarkMode();
 
 document.addEventListener("DOMContentLoaded", () => {
-  hljs.registerLanguage("python", hljsPython);
-  hljs.registerLanguage("shell", hljsShell);
-  hljs.registerLanguage("yaml", hljsYaml);
-  hljs.highlightAll();
   registerServiceWorker();
   darkMode.addToggleListener();
 });